// packages/github-mcp-server/src/common/types.ts
export const GithubFileContentLinks = z.object({
	self: z.string(),
	git: z.string().nullable(),
	html: z.string().nullable(),
});

export const GitHubFileContentSchema = z.object({
	name: z.string(),
	path: z.string(),
	sha: z.string(),
	size: z.number(),
	url: z.string(),
	html_url: z.string(),
	git_url: z.string(),
	download_url: z.string(),
	type: z.string(),
	content: z.string().optional(),
	encoding: z.string().optional(),
	_links: GithubFileContentLinks,
});

export const GitHubDirectoryContentSchema = z.object({
	type: z.string(),
	size: z.number(),
	name: z.string(),
	path: z.string(),
	sha: z.string(),
	url: z.string(),
	git_url: z.string(),
	html_url: z.string(),
	download_url: z.string().nullable(),
});

export const GitHubContentSchema = z.union([
	GitHubFileContentSchema,
	z.array(GitHubDirectoryContentSchema),
]);